respond to the PR comments
