+++ app/server-test/src/test/scala/org/bitcoins/server/DLCRoutesSpec.scala
@@ -2,7 +2,7 @@ package org.bitcoins.server
 
 import akka.http.scaladsl.model.ContentTypes
 import akka.http.scaladsl.testkit.ScalatestRouteTest
-import org.bitcoins.core.api.dlc.wallet.db.DLCContactDb
+import org.bitcoins.core.api.dlc.wallet.db.{DLCContactDb, DLCContactMappingDb}
 import org.bitcoins.core.currency.{Bitcoins, Satoshis}
 import org.bitcoins.core.protocol.dlc.models.ContractInfo
 import org.bitcoins.core.protocol.tlv.OracleAnnouncementTLV
@@ -211,7 +211,11 @@ class DLCRoutesSpec
       (mockWallet
         .addDLCContactMapping(_: Sha256Digest, _: InetSocketAddress))
         .expects(Sha256Digest.empty, expected.address)
-        .returning(Future.unit)
+        .returning(
+          Future.successful(
+            DLCContactMappingDb(Sha256Digest.empty,
+                                InetSocketAddress.createUnresolved("127.0.0.1",
+                                                                   1))))
 
       val args =
         ujson.Arr(ujson.Str(Sha256Digest.empty.hex), ujson.Str(address))
@@ -221,7 +225,8 @@ class DLCRoutesSpec
 
       Post() ~> route ~> check {
         assert(contentType == ContentTypes.`application/json`)
-        assert(responseAs[String] == s"""{"result":"ok","error":null}""")
+        assert(
+          responseAs[String] == s"""{"result":{"dlcId":"0000000000000000000000000000000000000000000000000000000000000000","contactId":"127.0.0.1:1"},"error":null}""")
       }
     }
