+++ dlc-wallet/src/main/scala/org/bitcoins/dlc/wallet/models/DLCContactMappingDAO.scala
@@ -1,6 +1,11 @@
 package org.bitcoins.dlc.wallet.models
 
-import org.bitcoins.core.api.dlc.wallet.db.{DLCContactDb, DLCDb}
+import org.bitcoins.core.api.dlc.wallet.db.{
+  DLCContactDb,
+  DLCContactMapping,
+  DLCContactMappingDb,
+  DLCDb
+}
 import org.bitcoins.crypto.Sha256Digest
 import org.bitcoins.db.{CRUD, SlickUtil}
 import org.bitcoins.dlc.wallet.DLCAppConfig
@@ -8,18 +13,6 @@ import slick.lifted.ForeignKeyQuery
 
 import java.net.InetSocketAddress
 import scala.concurrent.{ExecutionContext, Future}
-
-case class DLCContactMapping(dlc: DLCDb, contact: Option[DLCContactDb]) {
-
-  def dlcId: Sha256Digest = dlc.dlcId
-
-  def toDLCContactMappingDb: Option[DLCContactMappingDb] =
-    contact.map(c => DLCContactMappingDb(dlc.dlcId, c.address))
-}
-
-case class DLCContactMappingDb(
-    dlcId: Sha256Digest,
-    contactId: InetSocketAddress)
 
 case class DLCContactMappingDAO()(implicit
     override val ec: ExecutionContext,
